requirements.txt
google-genai
streamlit
python-dotenv
pydantic
requests
rich